// dashboard/src/app/types.ts
export interface User {
  id: string;
  firstName: string;
  lastName: string;
  avatar: string;
  email: string;
  role: string;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface Invitation {
  _id: string;
  teamId: number;
  userId: number;
  teamName: string;
  role: string;
  status: string;
  created_at: Date;
  updated_at: Date;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  start_date: Date;
  end_date: Date;
  status: string;
  owner_id: string;
  teams: Team[];
  tasks: Task[];
  created_at: Date;
  updated_at: Date;
}

export interface TeamMember {
  id: number;
  team_id: number;
  user_id: string;
  role: string
}

export interface Team {
  id: number;
  name: string;
  description: string;
  members: TeamMember[];
  created_at: Date;
  updated_at: Date;
  projects: Project[] | null;
}

export interface Task {
  _id: string;
  title: string;
  description: string;
  status: string;
  priority: string;
  project_id: number;
  team_id: number | null;
  assignee_id: string | null;
  creator_id: string;
  due_date: Date | null;
  estimated_hours: number | null;
  actual_hours: number | null;
  label: string | null;
  attachments: Attachment[];
  comments: Comment[];
  created_at: Date;
  updated_at: Date;
}

export interface Attachment {
  // Backend attachment properties
  _id?: string;
  task_id?: string;
  attachment_type?: string;
  attachment_url?: string;
  original_filename?: string;
  file_size?: number;
  uploaded_by: string;
  status?: string;
  created_at?: string;
  updated_at?: string;
  
  // Form upload properties
  file?: File;
  preview?: string;
}

export interface Comment {
  _id: string;
  task_id: string;
  project_id: number;
  creator_id: string;
  comment: string;
  created_at: Date;
  updated_at: Date;
}
// ...
export interface Plan {
  id: number;
  name: string;
  description: string;
  price: number;
  special_title?: string;
  is_active: boolean;
  billing_cycle: string;
  projects: number;
  teams: number;
  chat: boolean;
  priority: boolean;
  analytics: boolean;
  security: boolean;
}

export interface Invoice {
  id: number;
  subscription_id: number;
  stripe_invoice_id: string;
  stripe_payment_intent_id: string;
  amount: number;
  status: string;
  due_date: Date;
  paid_at: Date;
  invoice_number: string;
  subscription?: Subscription;
}

export interface Subscription {
  id: number;
  user_id: string;
  plan_id: number;
  start_date: string;
  end_date: string;
  status: string;
  stripe_subscription_id: string | null;
  stripe_customer_id: string | null;
  auto_renew: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  plan?: Plan;
}